+++ python_antenna_pattern/config.py
@@ -1,15 +1,11 @@
 ##### Default config parameters
 
 VERBOSE=False
-C250_FLAG=False
-C250_ROTATION_OFFSET=90
-MSIV_FLAG=True
-MSIV_ROTATION_OFFSET=90
+HOR_ROTATION_OFFSET=0   # Use 90 for C250 planet files where the angle is rotated by 90
+VER_ROTATION_OFFSET=90  # Use 90 for showing like most mobile antenna vendors draws
 MAX_GAIN_CLIPPING=30
 
-SIMULATE_FLAG=False
-SIMULATE_SAVE=False
-SIMULATE_TILT=5
+SIMULATE_SAVE=True
 MERGE_SAVE=True
 
 COLOR_3DB="gray"